spring/rest-sushi-shop/src/main/java/edu/api/ProductController.java
package edu.api;

import edu.domain.entities.Product;
import edu.domain.services.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/product")
public class ProductController {

    @Autowired
    private ProductService productService;

    @GetMapping("/{id}")
    public ResponseEntity<Product> findProductByID(@PathVariable Long id) {
        Product product = productService.findProductByID(id);
        return ResponseEntity.ok(product);
    }

}
